Revert "hooks"
