+++ src/components/SingleTrip.js
@@ -1,20 +1,33 @@
 import "../styles/SingleTrip.css";
+import { getSingleTripFromLocalStorage } from "../services/tripsStorage";
+import { useEffect, useState } from "react";
 import { useParams, Link } from "react-router-dom";
 import TripDetails from "./TripDetails";
 import Carousel from "./Carousel";
 import Weather from "./Weather";
 import GoBackButton from "./GoBackButton";
-import useSingleTrip from "../hooks/useSingleTrip";
 
 function SingleTrip() {
-  const [singleTrip] = useSingleTrip({});
+  const [singleTrip, setSingleTrip] = useState({});
   const { id } = useParams();
-  const [firstYear] = useSingleTrip("");
-  const [firstMonth] = useSingleTrip("");
-  const [firstDay] = useSingleTrip("");
-  const [secondYear] = useSingleTrip("");
-  const [secondMonth] = useSingleTrip("");
-  const [secondDay] = useSingleTrip("");
+  const [firstYear, setFirstYear] = useState("");
+  const [firstMonth, setFirstMonth] = useState("");
+  const [firstDay, setFirstDay] = useState("");
+  const [secondYear, setSecondYear] = useState("");
+  const [secondMonth, setSecondMonth] = useState("");
+  const [secondDay, setSecondDay] = useState("");
+
+  useEffect(() => {
+    const myTrip = getSingleTripFromLocalStorage(id);
+    setSingleTrip(myTrip);
+
+    setFirstYear(myTrip.dates[0].year);
+    setFirstMonth(myTrip.dates[0].month.name);
+    setFirstDay(myTrip.dates[0].day);
+    setSecondYear(myTrip.dates[1].year);
+    setSecondMonth(myTrip.dates[1].month.name);
+    setSecondDay(myTrip.dates[1].day);
+  }, [id]);
 
   return (
     <div className="singleTrip" key={id}>
