+++ src/components/Trips.js
@@ -3,12 +3,17 @@ import {
   getTripsFromLocalStorage,
   removeTripFromLocalStorage,
 } from "../services/tripsStorage";
+import { useEffect, useState } from "react";
 import NoTripsPlaceholder from "./NoTripsPlaceholder";
 import TripCard from "./TripCard";
-import useAllTrips from "../hooks/useAllTrips";
 
 function Trips() {
-  const [trips, setTrips] = useAllTrips([]);
+  const [trips, setTrips] = useState([]);
+
+  useEffect(() => {
+    const trip = getTripsFromLocalStorage("tripData");
+    setTrips(trip);
+  }, []);
 
   function handleRemoveTrip(trip) {
     const confirm = window.confirm("Do you really want to remove the trip?");
