--- src/hooks/useAllTrips.js
@@ -1,13 +1,0 @@
-import { useEffect, useState } from "react";
-import { getTripsFromLocalStorage } from "../services/tripsStorage";
-
-function useAllTrips() {
-  const [trip, setTrip] = useState([]);
-
-  useEffect(() => {
-    const trip = getTripsFromLocalStorage("tripData");
-    setTrip(trip);
-  }, []);
-  return [trip, setTrip];
-}
-export default useAllTrips;
